[Fix] Program 출시 전 마지막 점검
포탈 닫고, 몬스터 처치해야 다음 스테이지 갈 수 있음 & 오류 수정

+++ Unity/Assets/Script/Client/Database/Enemy/EnemySO.cs
@@ -14,7 +14,7 @@ public class EnemySO : ScriptableObject
     public float monsterHp;
     public float monstercurrHp;
     public float monsterAttack;
-    public int monsterAttackInterval;
+    public float monsterAttackInterval;
 
 
     public override string ToString()
